requirements.txt
SQLAlchemy~=2.0.17
fastapi~=0.99.1
passlib~=1.7.4
pydantic~=1.10.11
python-dotenv~=1.0.0
python-jose~=3.3.0
python-multipart~=0.0.6
uvicorn~=0.22.0
celery~=5.3.1
torch~=2.0.1
numpy~=1.25.1